Update GPSR.tex
Two points that I would like to see different:
1.: About the type of actions.
Last year at the WC only the action 'Bring an object from location1 to location2" was possible and that give already lots of troubles to most teams. Now there are lots of new action examples introduced which makes it even harder to understand all the possible actions. Some of them are quite unclear and therefore I propose some adjustments in the file.

Remarks:
Example-action1: Go to the kitchen, find a person and follow her (missing object-interaction part).
Remark on example-action1: The last part 'follow her', until the time ends..? or until she says stop? Very unclear.

Example-action2: Go to the dinner-table, grasp the crackers, and take them to the TV.
Remark on example-action2: Should the robot place the crackers on the TV in this example? Very unlikely that this must happen in real life. Since when does a TV like to have some crackers?

Example-action3: Offer a drink to the person at the door.
Remark on example-action3: I assume that the robot needs to know where the drinks are stored. This information will be given? 

2.: In the score system the robot has only one attempt to understand the operator correctly. Otherwise no points. To my opinion, a robot should have more chances to score points, since speech recognition is such an important task in this challenge.