Include the Filling the Dishwasher test (Stage I) in the rulebook
A new test is introduced in Stage 1 for OPL and DSPL (mostly
manipulation based), the Filling the Dishwasher test, granting
up to 350 points for allocating up to 10 objects in a dishwasher.

General rules are modified to mention the Dishwasher in the arena
specification, as well as the utensils to be used in this test,
for which colorful examples have been chosen to lower the gap in
object recognition in this first try.

Also, the test has been designed to be compliant with
issues #358, #353, and with P&G request