Include the Dishwasher Master test (Stage II) in the rulebook
A new test is introduced in Stage 2 for OPL and DSPL (mostly
manipulation based), the Dishwasher Master test, granting
up to 600 points for allocating up to 10 objects in a dishwasher.
This test considers also opening and closing the dishwasher as
well as placing the soap tab in the corresponding compartment.

General rules are modified to mention the Dishwasher in the arena
specification, as well as the utensils to be used in this test,
for which colorful examples have been chosen to lower the gap in
object recognition in this first try.

Also, the test has been designed to be compliant with
issues #358, #353, and with P&G request