Include the Cleaning the Toilet test (Stage I) in the rulebook
A new test is introduced in Stage 1 for OPL and DSPL (mostly
manipulation based), the Cleaning the Toilet test, granting
up to 350 points for cleaning up chocolate syrup from a toilet
using a brush or sponge.

General rules are modified to mention the Toilet in the arena
specification, as well as the brush, which can be provided by
the team.

Also, the test has been designed to be compliant with issue #358