Add Official Standard Laptop specs and rules (compulsory for DSPL)
This patch adds to the rulebook the "Official Standard Laptop for DSPL"
section. This section makes mandatory the use of the OSL in DSPL, as well
as restricting the use of any hardware attached to the OSL and allowing
Referees/TC/OC to run random checks.

Robot Inspection is modified to include the inspection of the OSL

Changes are compliant with what was discussed and decided in #366 and #368