Rework and balance Storing Groceries Test
The test as described promoted no manipulation since most points were granted for object recognition and not for solving a task (#353). To solve this issue, now points are granted per moved object and the PDF report is only for validation purposes. In addition, the number of points granted per object increases with every move, promoting speed. The description of the task was changed accordingly. Scoring system redesigned focusing on actually solving the task and addressing #358.

Changes:
- Setup:
  - Test can take place inside or outside the arena
    (explicit)
  - Buffed other setup items and removed redundant references
- Task:
  - Stacking objects is allowed (*new*).
  - Finding the table, inspecting the cupboard,
    and opening the door can be done in any order.
- Additional Rules and remarks
  - Removed PDF report.
  - Added timing section.
- Moved PDF report instructions to its own section.
  - Report only considered if the robot moved at
    least on object.
  - Removed final state (same as all no initial states).
- PDF report is OPTIONAL.
- Scoresheet
  - Simplified timing
  - Reworked
  - Incremental scoring per object placed