REMOVE: Set Table
This task is admittedly one of the harder cuts to make.
But we must make the RoboCup@Home Competition clearer.
Perhaps it will be added back in the future, or an extension of test.
The robotics related components of this test are covered in other tests.

Related to #685
- 10 teams opted into this task before Sydney
- 1 teams attempted this task in Sydney
- 1 teams scored on this task in Sydney

Similar to other removal PRs: #682 #688 #689