[tasks+scoresheets/SmoothieChef] Updated scores; added penalty for incorrect fruit order
The score update is so that, for consistency, we have 1200 main points
and 800 bonus points.

According to the original formulation, points were only scored if the
fruits were placed in the blender in the correct order, which was too
harsh in my opinion. In the updated formulation, teams also get points
if they place the fruits in the incorrect order, but the score is reduced.