[Receptionist] updates for the challenge (#930)
# Proposed Solution for issue #916 
with @sunava 
Changes proposed in this pull request:
- the robot shows both guests where the drinks are and detects if their
favorite drink is available
- There are passive guests in the background
- Points for finding one hobby/interest of each guest
- At the end, only the guests get introduced 
- optional goal 'describing the first guest' changes. describe first
guest to second guest **before** reaching the living room (updated
10.01)
- guest paces from left to right during conversation with robot (update
10.01)
-  ~~Added optional goal picking up a drink~~ (updated 10.01)
- Added optional goal for finding similar interest between guests or
guest and host
- added a procedural guide for quicker and easier understanding of the
challenge
- Changed scoresheet accordingly

## Updated procedure summarized
Both guests arrive separately. The robot either opens the door for the
guest or
waits for them at the starting point. It greets the guest and asks for
their name.
The robot then guides the guest to the beverage area, where it asks for
their
favorite drink and checks if the drink is available on the table. The
robot gathers
one interest of the guest at a freely chosen moment.
After showing the guest the beverage area, the robot escorts them to the
living
room and offers a free seat. Once both guests are seated, the robot
introduces
them to each other.

## Other comments
- ~~I would suggest making it an obligation to log what the robot is
saying.
This makes the process easier for the referee in a noisy arena to check
the output quickly after the challenge~~

- The passive guests and the host could be spectators if it is not too
much effort.

---------

Co-authored-by: Tiago Ribeiro <tiagoribeiro80@gmail.com>
Co-authored-by: Leroy Rügemer <lruegeme@techfak.uni-bielefeld.de>