[GO feedback] SG: reduce point for opening the second cabinet door (#953)
## Description

During the German open FAQ teams we specified that asking the referee to
open one door still enables the robot to score opening the second door.
As the first door is much harder to open (need to use handle or rims )
than the second (possible grasp behind) It should reward a different
number of points.

---------

Co-authored-by: Matthijs van der Burgh <MatthijsBurgh@outlook.com>