Various changes to phrasing, labels, command invocations (#981)
various small changes:
- Latex: ensure labels point to the correct location
- Latex: Terminate commands with {}
- various typo fixes
- changes to phrasing
- Fix: remove allowing two robots during final - not allowed with new
final, we missed this

## Other comments
I read the rulebook and tried to improve the phrasing along the way to
make the rules a little more clear in various places.

sorry for the big changeset.
I can split this up (after competition) if it is too much to perform a
thorough review.