First draft of Restaurant 2026 (#1000)
## Description
Updates for Restaurant 2026
To encourage teams to use barman rather than being stuck in the middle
of the restaurant.
- increased the time and count to instruct the Barman from one minute
for the task to two minutes per order.
- add some examples of instructions
- reduced the penalty score of "Being guided to a table"
- add penalty "Asking for directional confirmation" due to no awareness

I didn't include the barman's guidance/pointing procedure that
automatically executed after 1 minute.
Because it will not change the robot's behavior and only make it look
worse.